% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_grand_mean.R
\name{center_grand_mean}
\alias{center_grand_mean}
\title{Center with respect to grand mean}
\usage{
center_grand_mean(data, cols, keep_original = TRUE)
}
\arguments{
\item{data}{dataframe}

\item{cols}{vector or tidyselect syntax or helpers. column(s) that need to be centered. Recommend using where(is.numeric) to exclude changing factors.}

\item{keep_original}{default is `FALSE`. Set to `TRUE` to keep original columns}
}
\value{
return a dataframe with the columns centered (replace existing columns)
}
\description{
Center all columns with respect to the grand mean
}
\examples{
center_grand_mean(iris,where(is.numeric))

}
