% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_score.R
\name{composite_score}
\alias{composite_score}
\title{Composite column}
\usage{
composite_score(
  data,
  cols = tidyselect::everything(),
  na.rm = FALSE,
  composite_col_name = "composited_column"
)
}
\arguments{
\item{data}{dataframe}

\item{cols}{vector or tidyselect syntax or helpers. column(s) that need to be composite}

\item{na.rm}{ignore NA}

\item{composite_col_name}{character. default as 'composite_column'. the column name of the composite column}
}
\value{
return a dataframe with a new column with the composite score
}
\description{
The function will perform a rowise aggregation which then divided by the total number of columns. It ignores rows with any NA value (i.e, will use listwise deletion)
}
\examples{
test_df = data.frame(col1 = c(1,2,3,4),col2 = c(1,2,3,4), col3 = c(1,2,NA,4))
composite_df = composite_score(data = test_df)


}
