% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_coding.R
\name{dummy_coding}
\alias{dummy_coding}
\title{Dummy Coding}
\usage{
dummy_coding(data, cols)
}
\arguments{
\item{data}{data.frame object}

\item{cols}{Columns that need to be dummy-coded See `dplyr::dplyr_tidy_select` for available options.}
}
\value{
An object of the same type as .data. The output has the following properties:
1. Columns from .data will be preserved.
2. Columns that are dummy-coded.
}
\description{
Create dummy-coded columns, supporting tidyselect syntax to process multiple columns simultaneously.
}
\examples{
dummy_coding(iris,Species)

}
