% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlbook_data.R
\docType{data}
\name{mlbook_data}
\alias{mlbook_data}
\title{mlbook_data}
\format{
A data frame with 3758 rows and 34 variables:
\describe{
  \item{schoolnr}{School ID}
  \item{pupilNR_new}{Student Identifier (Level 1 units)}
  \item{langPOST}{Student language score}
  \item{ses}{Student socioeconomic score, grand-mean centered (in points, M = 0))}
  \item{IQ_verb}{Student verbal IQ, grand-mean centered (in points, M = 0)}
  \item{sex}{Student binary gender, 1 = female, 0 = not female}
  \item{Minority}{Student minority status, 1 = minoritized, 0 = not minoritized}
  \item{denomina}{School-level religious denominations, 5 categories}
  \item{female_dum}{Dummy coded sex}
  \item{female_eff}{Effect-coded sex}
  \item{female_CMC}{Group-mean-centered of female_eff}
  \item{fempct_agg}{Aggregated mean female_dum for each school}
  \item{Zfempct_agg}{Z-scored aggregated mean female_dum for each school}
  \item{ses_CMC}{Group-mean-centered SES}
  \item{Zses_CMC}{Z-scored group-mean-centered SES}
  \item{ses_agg}{Aggregated mean SES for each school}
  \item{Zses_agg}{Z-scored aggregated mean SES for each school}
}
}
\source{
\url{https://www.stats.ox.ac.uk/~snijders/mlbook.htm}
}
\usage{
mlbook_data
}
\description{
Classic data-set from Snijders, Tom A.B., and Bosker, Roel J. Multilevel Analysis: An Introduction to Basic and Advanced Multilevel Modeling, second edition.
}
\keyword{datasets}
