% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_all.R
\name{render_all}
\alias{render_all}
\title{RMarkdown Render}
\usage{
render_all(dir, render_exist = T)
}
\arguments{
\item{dir}{directory path}

\item{render_exist}{render existed pdf files or not}
}
\description{
It will render all RMarkdown files in the directory.
}
