% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_non_missing_values.R
\name{summarize_non_missing_values}
\alias{summarize_non_missing_values}
\title{Count number of non-missing / non-NA values}
\usage{
summarize_non_missing_values(
  data,
  cols = dplyr::everything(),
  group = NULL,
  print = F
)
}
\arguments{
\item{data}{dataframe}

\item{cols}{vector or tidyselect syntax or helpers. default is all columns}

\item{group}{character. pass to group_by}

\item{print}{return the original data, and print the summary}
}
\value{
return a dataframe with the number of non-NA values
}
\description{
It counts the number of non-missing / non-NA values. You can pass everything() to cols in order to get all the missing value for all columns
}
\examples{
test_df = data.frame(col1 = c(1,2,3),col2 = c(1,NA,3),col3 = c(1,2,NA))
summarize_non_missing_values(test_df,everything())

}
