% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_scored_group_mean.R
\name{z_scored_group_mean}
\alias{z_scored_group_mean}
\title{Z scored with with respect to the group mean}
\usage{
z_scored_group_mean(data, cols, group, keep_original = TRUE)
}
\arguments{
\item{data}{A data.frame or a data.frame extension (e.g. a tibble).}

\item{cols}{Columns that need to be centered. See `dplyr::dplyr_tidy_select` for available options.}

\item{group}{the grouping variable. If you need to pass multiple group variables, try to use quos(). Passing multiple group variables is not tested.}

\item{keep_original}{default is `FALSE`. Set to `TRUE` to keep original columns}
}
\value{
return a dataframe with a group-mean centered columns that are z-scored with respect to the grand mean.
}
\description{
This function will compute group-mean-centered scores, and then z-scored the group-mean-centered scores with respect to the grand mean.
}
\examples{
z_scored_group_mean(iris, dplyr::ends_with("Petal.Width"), "Species")
}
