% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_scored_mlm.R
\name{z_scored_mlm}
\alias{z_scored_mlm}
\title{Z-scored for multilevel analyses
It will group mean center and z-scored the L1 score and create a mean score for each L2 group.}
\usage{
z_scored_mlm(data, cols, group, keep_original = TRUE)
}
\arguments{
\item{data}{dataframe}

\item{cols}{vector or tidyselect syntax or helpers. column(s) that need to be centered}

\item{group}{the grouping variable. Must be character}

\item{keep_original}{default is `FALSE`. Set to `TRUE` to keep original columns}
}
\value{
return a dataframe with the columns z-scored (replace existing columns)
}
\description{
Z-scored for multilevel analyses
It will group mean center and z-scored the L1 score and create a mean score for each L2 group.
}
\examples{
z_scored_mlm(iris,dplyr::ends_with('Length'),group = 'Species')

}
