% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_group_mean.R
\name{center_group_mean}
\alias{center_group_mean}
\title{Center with respect to group mean}
\usage{
center_group_mean(data, cols, group)
}
\arguments{
\item{data}{data frame.}

\item{cols}{vector or tidyselect syntax or helpers. column(s) that need to be centered.  Recommend using where(is.numeric) to exclude changing factors.}

\item{group}{character. grouping variable}
}
\value{
return a data frame with the columns centered (replace existing columns)
}
\description{
Center all columns with respect to the group mean.
}
\examples{
center_group_mean(iris,where(is.numeric), group = Species)

}
