% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_group_mean.R
\name{center_group_mean}
\alias{center_group_mean}
\title{Center with respect to group mean}
\usage{
center_group_mean(data, cols, group)
}
\arguments{
\item{data}{dataframe}

\item{cols}{vector or tidyselect syntax or helpers. column(s) that need to be centered.  Recommend using where(is.numeric) to exclude changing factors.}

\item{group}{the grouping variable. If you need to pass multiple group variables, try to use quos(). Passing multiple group variables is not tested.}
}
\value{
return a dataframe with the columns centered (replace existing columns)
}
\description{
Center all columns with respect to the group mean. Currently, it will drop the group column (Will try to fix this in the next version)
}
\examples{
center_group_mean(iris,where(is.numeric), group = Species)

}
