% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_mlm.R
\name{center_mlm}
\alias{center_mlm}
\title{Centering for multilevel analyses
It will group mean center the L1 score and create a mean score for each L2 group.}
\usage{
center_mlm(data, cols, group, keep_original = TRUE)
}
\arguments{
\item{data}{dataframe}

\item{cols}{vector or tidyselect syntax or helpers. column(s) that need to be centered}

\item{group}{the grouping variable. Must be character}

\item{keep_original}{default is `FALSE`. Set to `TRUE` to keep original columns}
}
\value{
return a dataframe with the columns z-scored (replace existing columns)
}
\description{
Centering for multilevel analyses
It will group mean center the L1 score and create a mean score for each L2 group.
}
\examples{
center_mlm(iris,ends_with('Length',group = 'Species')

}
