% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_coding.R
\name{dummy_coding}
\alias{dummy_coding}
\title{Dummy Coding
Create dummy coding columns}
\usage{
dummy_coding(data, cols)
}
\arguments{
\item{data}{data.frame object}

\item{cols}{vector or tidyselect syntax or helpers. column(s) that need to be re-coded}
}
\value{
return a data.frame with factors being dummy coded
}
\description{
Dummy Coding
Create dummy coding columns
}
\examples{
dummy_coding(iris,Species)

}
