% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_coding.R
\name{effect_coding}
\alias{effect_coding}
\title{Effect Coding}
\usage{
effect_coding(data, cols, factor = FALSE)
}
\arguments{
\item{data}{A data.frame or a data.frame extension (e.g. a tibble).}

\item{cols}{Columns that need to be effect-coded. See `dplyr::dplyr_tidy_select` for available options.}

\item{factor}{The default is `FALSE`. If factor is set to `TRUE`, this function returns a tibble with effect-coded factors. If factor is set to `FALSE`, this function returns a tibble with effect-coded columns.}
}
\value{
An object of the same type as .data. The output has the following properties:
1. Columns from .data will be preserved.
2. Columns that are effect-coded.
}
\description{
Create effect-coded columns, supporting tidyselect syntax to process multiple columns simultaneously.
}
\examples{
effect_coding(iris,Species)

}
