% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_item.R
\name{recode_item}
\alias{recode_item}
\title{Recode items of a dataframe (support reverse code)}
\usage{
recode_item(
  data,
  cols,
  code_from = NULL,
  code_to = NULL,
  reverse_code = FALSE,
  retain_code = "all"
)
}
\arguments{
\item{data}{a dataframe}

\item{cols}{vector or tidyselect syntax or helpers. column(s) that need to be recoded}

\item{code_from}{vector. the order must match with vector for code_to}

\item{code_to}{vector. the order must match with vector for code_from}

\item{reverse_code}{logical. Default as F. If T, it will use psych::reverse_code}

\item{retain_code}{vector. the code to be retained. default is everything.}
}
\description{
Recode items of a dataframe (support reverse code)
}
\examples{
test_df = data.frame(test_col = 1:5)
recode_item(test_df, cols = test_col, reverse_code = TRUE, retain_code = 1:4)

}
