% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_non_missing_values.R
\name{summarize_non_missing_values}
\alias{summarize_non_missing_values}
\title{Count number of non-missing / non-NA values}
\usage{
summarize_non_missing_values(
  data,
  cols = dplyr::everything(),
  group = NULL,
  print = FALSE
)
}
\arguments{
\item{data}{data frame.}

\item{cols}{vector or tidyselect syntax or helpers. default is all columns.}

\item{group}{character. If specify, the return data frame would show the number of missing values of each column by group.}

\item{print}{printing the return data frame.}
}
\value{
return a data frame with the number of non-NA values of each column.
}
\description{
It counts the number of non-missing / non-NA values. You can pass everything() to cols in order to get all the missing value for all columns
}
\examples{
test_df = data.frame(col1 = c(1,2,3),col2 = c(1,NA,3),col3 = c(1,2,NA))
summarize_non_missing_values(test_df,everything())

}
