% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_scored_grand_mean.R
\name{z_scored_grand_mean}
\alias{z_scored_grand_mean}
\title{Z scored with with respect to the grand mean}
\usage{
z_scored_grand_mean(data, cols)
}
\arguments{
\item{data}{dataframe}

\item{cols}{vector or tidyselect syntax or helpers. column(s) that need to be centered}
}
\value{
return a dataframe with the columns z-scored (replace existing columns)
}
\description{
Z-scored that uses the grand mean in the z-score formula
}
\examples{
z_scored_grand_mean(iris,where(is.numeric))

}
