% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_scored_group_mean.R
\name{z_scored_group_mean}
\alias{z_scored_group_mean}
\title{Z scored with with respect to the group mean}
\usage{
z_scored_group_mean(data, cols, group, keep_original = TRUE)
}
\arguments{
\item{data}{dataframe}

\item{cols}{vector or tidyselect syntax or helpers. column(s) that need to be centered}

\item{group}{the grouping variable. If you need to pass multiple group variables, try to use quos(). Passing multiple group variables is not tested.}
}
\value{
return a dataframe with the columns z-scored (replace existing columns)
}
\description{
Z-scored that uses the group mean in the z-score formula. Currently, it will drop the group column (Will try to fix this in the next version)
}
\examples{
z_scored_group_mean(iris, where(is.numeric), group = Species)
z_scored_group_mean(iris, 1:4, group = Species)

}
